% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_draws.R
\name{get_draws}
\alias{get_draws}
\title{Get Draws}
\usage{
get_draws(mean, sd, host = "localhost")
}
\arguments{
\item{host}{}
}
\description{
Get Draws
}
\examples{
\dontrun{
draws <- testShinyPlumber::get_draws(mean = 9, sd = 1)
}
}
